feat(parser): expose findESIIncludeTags for ESI pre-processing (#12)
We need to pre-process the ESI tags so that they play nicely with a standard HTML parser, which does not recognize self-closing tags. Instead of rolling our own regex-based ESI parser, we'd like to use `nodesi`. Therefore this change renames `findESIInclueTags` to `findESIIncludeTags` and exposes it as part of the `ESI` API.

see https://github.com/adobe/helix-pipeline/issues/317