feat(esi): Add support for `alt` attribute
The ESI spec has built-in support for fallbacks using the `alt` attribute. This change adds the required support. If an `alt` attribute is present and the `src` cannot be retrieved, another retrieval will be attempted, this time using the `alt` source. If this fails, too, the usual error handler will be invoked.

fixes #19