Add support for the alt attribute (#22)
* docs(license): clarify use of ISC license

So far, the license was mentioned only in `package.json`, so I'm adding a proper `LICENSE.txt` as well as links in `README.md`

as discussed in https://github.com/Schibsted-Tech-Polska/nodesi/pull/12#issuecomment-491293225

* Update LICENSE.txt

* fix(parser): normalize entities in URL before calling data provider

this change is using the external dependency `he`, which was already included in the dependency tree

fixes #17

* builld(dependencies): make he dependency explicit

this is adding an explicit dependency to `he`, in case `good-guy-http` decides to drop it from the dependency stack

* feat(esi): Add support for `alt` attribute

The ESI spec has built-in support for fallbacks using the `alt` attribute. This change adds the required support. If an `alt` attribute is present and the `src` cannot be retrieved, another retrieval will be attempted, this time using the `alt` source. If this fails, too, the usual error handler will be invoked.

fixes #19