feat: remove vulnerable libs
This requires a change of the http request lib. Now nodesi uses
a native fetch, or you can provide your own http client
implementation.

Previously nodesi used good-guy-http, which is no longer
maintained, so the change in the behavior is that nodesi
does not provide a cache out of the box.

good-guy-http came with the in-memory cache that respected
Cache-Control headers, but the new native fetch implementation
does not provide this feature.

BREAKING CHANGE: change in the default http request library